feat: add list view layout
* refactor!: add layout change buttons and move folder actions to nav
Now depending on the layout received from the server, the items are displayed differently.

Grid no longer contains the folder actions except for the 'go back' button.
The actions are moved to the nav bar.

Display a 'call to action' element when no folders or files are present in his drive.

 * feat: create row item component for list view

 * feat: create list view for folder items that displays them as a table

 * feat: add number of files in a folder for size in item card
 W.I.P. on the server side.

 * feat: configure getDrive and getFolder endpoints to get layout from query params
If no layout is detected then the default is grid

 * fix: change redirect to show the right page in file upload and update folder

 * feat: add url to res.locals

 * feat: add user agent middleware to set default layout based on device type

 * feat: add layout field to req session data prop

 * feat: add useragent express middleware to app.ts

 * feat: add layout types

 * build: add express-useragent dependency