feat: add shared folder functionality and make ejs views more flexible
 * feat: add shared folder file download

 * feat: add GET search shared folder endpoint

 * feat(flexible layout): implement factory method for shared folder view

 * feat(flexible layout): create SharedFolderViewModel and factory method

 * feat(flexible layout): implement the VM related to user search in drive controller

 * feat(flexible layout): implement the VM related to link creation in drive controller

 * feat(flexible layout): create LinkCreatedViewModel and factory method
Kinda testing out the waters with this one.
Get the feel for this kind of workflow.
This is for the more specific scenario when the user shares a folder and link is generated.

 * feat(flexible layout): create UserSearchResultsViewModel with factory method

 * feat(flexible layout): create BaseSearchResultsViewModel that inherits BaseDriveViewModel
This VM contains only the search and content layout related actions.
Meaning you cant create a folder or add a file when you are seeing search results.

 * refactor(flexible layout): enhance /drive page rendering with UserDriveViewModel

 * refactor(flexible layout): change ejs render logic to be mainly based on params
This is for the layouts and the drive.ejs and search.ejs views.
Now things are more flexible.

 * refactor(flexible layout): create a drive page main layout
The layout is based on the drive ejs view.
The difference is that it contains params for:
- navigation,
- sidebar
- main content
- and dialog

Params are HTML unescaped specifically for ejs partials and components.

 * feat: create UserDriveViewModel interface and factory function with defaults

 * feat: add BaseDriveViewModel interface and related types for drive functionality

 * refactor(flexible layout): move dialog related partials to their specific section
The copy-folder-link and share-folder-form partials will only appear in the /drive routes.
So they have been placed in partials/drive/

 * feat(flexible layout): add param for dropdown availability in card and row components

 * refactor(flexible layout): make the grid layout for drive have more general names
The current goal is moving the logic away from the partials.
The replacements are more param heavy for readability.

 * refactor(flexible layout): create an ejs sidebar partial for reuse

 * refactor(flexible layout): create an ejs navigation partial for reuse
Because the navigation is the same across multiple pages, it can be extracted.
The new partial has more parameters, but changing it is easier.
1 source of bugs and 1 source of change.

 * feat: update file view to conditionally render action buttons based on access type

 * feat: add GET shared folder file endpoint

 * feat: include shared folder owner to response locals

 * feat: add GET endpoints for shared folders and directories within

 * feat: create a helper for checking if a dir is part of a folder tree

 * feat: return error when shared folder links are no longer valid

 * feat: create 403 FORBIDDEN status error
This will be used for actions not allowed on public resources like the shared links

 * feat: add public router to app.ts

 * feat: create 410 GONE status error
This will be used for expired resources like the shared links

 * feat: implement folder link handler and public router for shared folders
Shared folder will be used all queries for '/public/:uuid/*', so a middleware sets it
That way the shared folder can be access in res.locals later

 * feat: add dialog for copy link that appears in drive on shared folder creation

 * feat: add 'copy link' section for link created for shared folder

 * feat: add button that shows form dialog for sharing folder

 * feat: create shared folder duration and access type form

 * feat: create dialog ejs component

 * feat: add folder sharing POST endpoint

 * feat: add shared folder table to prisma schema
A folder can have many shared folders.
When a folder is deleted, all shared folders are deleted too.

Shared folder has an expire date and access types